networkd/dhcp: implement RFC3396 for large DHCP options
A DHCP option is 255 bytes at most. However, some values may need more
data (e.g. Option 121 "Classless Static Routes"), for which the option
body is split in multiple option instances on the wire.

While the current implementation works well when the serializer splits
the option at logical boundaries (e.g. at the bounds of a whole route),
it fails as soon as it is split at arbitrary bytes boundaries.

This patch implements the aggregation of the option buffers in a single
large buffer as per the RFC, which is then fed to the actual parsing
functions.

Note: some options *require* support for concatenations, some of which
are already supported by networkd (opt 121 being a prime example).

+++ src/libsystemd-network/sd-dhcp-server.c
@@ -729,7 +729,7 @@ static int server_send_forcerenew(
                                     sizeof(DHCPMessage) + optoffset);
 }
 
-static int parse_request(uint8_t code, uint8_t len, const void *option, void *userdata) {
+static int parse_request(uint8_t code, size_t len, const void *option, void *userdata) {
         DHCPRequest *req = ASSERT_PTR(userdata);
         int r;
